+++ src/main/java/org/ojalgo/optimisation/linear/DenseTableau.java
@@ -169,7 +169,7 @@ final class DenseTableau extends SimplexTableau {
             }
         }
 
-        return retVal;
+        return retVal + this.getValueShift();
     }
 
     @Override